+++ modules/qna/src/backend/storage.ts
@@ -1,99 +1,11 @@
 import * as sdk from 'botpress/sdk'
-import fse from 'fs-extra'
 import _ from 'lodash'
-import mkdirp from 'mkdirp'
-import nanoid from 'nanoid/generate'
-import path from 'path'
-import tar from 'tar'
-import tmp from 'tmp'
 
 import { ImportArgs } from './api'
-import { Dic, Item } from './qna'
-import { ItemSchema, QnASchemaArray } from './validation'
-
-const safeId = (length = 10) => nanoid('1234567890abcdefghijklmnopqrsuvwxyz', length)
-const slugify = s => (s || '').toLowerCase().replace(/[^a-z0-9]/g, '_')
-
-const makeId = (item: Item) => {
-  const firstQuestion = item.questions[Object.keys(item.questions)[0]][0]
-  return `__qna__${safeId()}_${slugify(firstQuestion)
-    .replace(/^_+/, '')
-    .substring(0, 50)
-    .replace(/_+$/, '')}`
-}
-
-const normalizeQuestions = (questions: string[]) =>
-  questions
-    .map(q =>
-      q
-        .replace(/[\r\n]+/g, ' ')
-        .replace(/\s+/g, ' ')
-        .trim()
-    )
-    .filter(Boolean)
-
-interface Metadata {
-  answers: Dic<string[]>
-  contentAnswers: sdk.Content.All[]
-  enabled: boolean
-  lastModifiedOn: Date
-}
-
-type Intent = Omit<sdk.NLU.IntentDefinition, 'metadata'> & { metadata?: Metadata }
-
-const FLOW_FOLDER = 'flows'
-const MEDIA_FOLDER = 'media'
-const TEMP_INTENT_FILENAME = 'exportable.qna.intents.json'
-const INTENT_FILENAME = 'qna.intents.json'
-const QNA_PATH_TO_ZIP = path.join(FLOW_FOLDER, TEMP_INTENT_FILENAME)
-
-const toQnaFile = topicName => path.join(topicName, INTENT_FILENAME)
-const toZipFile = topicName => `${topicName}.qna.tar.gz`
-const serialize = (intents: Intent[]) => JSON.stringify(intents, undefined, 2)
-
-const keepEndPath = path =>
-  path
-    .split('/')
-    .slice(-2)
-    .join('/')
-
-const removeBotPrefix = (c: sdk.Content.All) => {
-  if ((c as sdk.Content.Carousel).items) {
-    const carousel = c as sdk.Content.Carousel
-    return {
-      ...carousel,
-      items: carousel.items.map(o => {
-        return { ...o, image: keepEndPath(o.image) }
-      })
-    }
-  } else if ((c as sdk.Content.Card | sdk.Content.Image).image) {
-    const image = c as sdk.Content.Image | sdk.Content.Card
-    return {
-      ...image,
-      image: keepEndPath(image.image)
-    }
-  } else {
-    return c
-  }
-}
-
-const addBotPath = (file, botId) => path.join(`/api/v1/bots/${botId}`, file)
-const addBotPrefix = (c: sdk.Content.All, botId: string) => {
-  if ((c as sdk.Content.Carousel).items) {
-    const carousel = c as sdk.Content.Carousel
-    return {
-      ...carousel,
-      items: carousel.items.map(o => {
-        return { ...o, image: addBotPath(o.image, botId) }
-      })
-    }
-  } else if ((c as sdk.Content.Card | sdk.Content.Image).image) {
-    const image = c as sdk.Content.Image | sdk.Content.Card
-    return { ...image, image: addBotPath(image.image, botId) }
-  } else {
-    return c
-  }
-}
+import { exportQuestionsToArchive, importArchivePerTopic } from './archive'
+import { Intent, Item } from './qna'
+import { FLOW_FOLDER, INTENT_FILENAME, makeId, normalizeQuestions, serialize, toQnaFile } from './utils'
+import { ItemSchema } from './validation'
 
 export default class Storage {
   constructor(private ghost: sdk.ScopedGhostService) {}
@@ -171,44 +83,8 @@ export default class Storage {
   }
 
   async exportPerTopic(topicName: string) {
-    const tmpDir = tmp.dirSync({ unsafeCleanup: true })
-    const zipName = path.join(tmpDir.name, toZipFile(topicName))
-    await mkdirp.sync(path.join(tmpDir.name, MEDIA_FOLDER))
-
     const jsonQnaForBot = await this.ghost.readFileAsObject<Intent[]>(FLOW_FOLDER, toQnaFile(topicName))
-    const jsonQnaBotAgnostic = jsonQnaForBot.map(i => {
-      const contentAnswers = [...(i.metadata?.contentAnswers ?? [])].map((c: sdk.Content.All) => removeBotPrefix(c))
-      const metadata = { ...i.metadata, contentAnswers }
-      return { ...i, metadata }
-    })
-
-    const medias = _.chain(jsonQnaBotAgnostic)
-      .flatMapDeep(item => item.metadata?.contentAnswers ?? [])
-      .flatMap(c => {
-        if ((c as sdk.Content.Carousel).items) {
-          const carousel = c as sdk.Content.Carousel
-          return carousel.items
-        } else {
-          return [c]
-        }
-      })
-      .map(c => _.get(c, 'image', undefined))
-      .filter(c => c)
-      .value()
-
-    await mkdirp.sync(path.join(tmpDir.name, path.dirname(QNA_PATH_TO_ZIP)))
-    await fse.writeFile(path.join(tmpDir.name, QNA_PATH_TO_ZIP), Buffer.from(serialize(jsonQnaBotAgnostic)))
-    await Promise.all(
-      medias.map(
-        async mediaFile =>
-          await fse.writeFile(path.resolve(tmpDir.name, mediaFile), await this.ghost.readFileAsBuffer('/', mediaFile))
-      )
-    )
-
-    await tar.create({ cwd: tmpDir.name, file: zipName, portable: true, gzip: true }, [...medias, QNA_PATH_TO_ZIP])
-    const zipBuffer = await fse.readFile(zipName)
-    tmp.setGracefulCleanup()
-    return zipBuffer
+    return exportQuestionsToArchive(jsonQnaForBot, topicName, this.ghost)
   }
 
   async moveToAnotherTopic(sourceTopic: string, targetTopic: string) {
@@ -227,33 +103,6 @@ export default class Storage {
   }
 
   async importArchivePerTopic(importArgs: ImportArgs) {
-    const { topicName, botId, zipFile } = importArgs
-
-    const tmpDir = tmp.dirSync({ unsafeCleanup: true })
-    await fse.writeFile(path.join(tmpDir.name, toZipFile(topicName)), zipFile)
-    await tar.extract({ cwd: tmpDir.name, file: path.join(tmpDir.name, toZipFile(topicName)) })
-
-    let qnasBotAgnostic: Intent[] = JSON.parse(
-      await fse.readFile(path.join(tmpDir.name, FLOW_FOLDER, TEMP_INTENT_FILENAME), 'utf8')
-    )
-    qnasBotAgnostic = await QnASchemaArray.validate(qnasBotAgnostic)
-
-    const qnaForBot: Intent[] = qnasBotAgnostic.map(i => {
-      const newContent = [...i.metadata.contentAnswers].map(c => addBotPrefix(c, botId))
-      const newMetadata = { ...i.metadata, contentAnswers: newContent }
-      return { ...i, contexts: [topicName], filename: path.join(topicName, INTENT_FILENAME), metadata: newMetadata }
-    })
-
-    await this.ghost.upsertFile(FLOW_FOLDER, path.join(topicName, INTENT_FILENAME), serialize(qnaForBot))
-
-    if (await fse.pathExists(path.join(tmpDir.name, MEDIA_FOLDER))) {
-      for (const mediaFile of await fse.readdir(path.join(tmpDir.name, MEDIA_FOLDER))) {
-        await this.ghost.upsertFile(
-          MEDIA_FOLDER,
-          mediaFile,
-          await fse.readFile(path.join(tmpDir.name, MEDIA_FOLDER, mediaFile))
-        )
-      }
-    }
+    return importArchivePerTopic(importArgs, this.ghost)
   }
 }
