+++ modules/qna/src/backend/utils.ts
@@ -1,10 +1,38 @@
 import * as sdk from 'botpress/sdk'
 import _ from 'lodash'
+import nanoid from 'nanoid/generate'
+import path from 'path'
 
-import { Item } from './qna'
+import { Intent, Item } from './qna'
+
+const slugify = s => (s || '').toLowerCase().replace(/[^a-z0-9]/g, '_')
 
 export const QNA_MIN_QUESTIONS = 3
 export const QNA_MIN_ANSWERS = 1
+export const INTENT_FILENAME = 'qna.intents.json'
+export const FLOW_FOLDER = 'flows'
+
+export const toQnaFile = topicName => path.join(topicName, INTENT_FILENAME)
+export const serialize = (intents: Intent[]) => JSON.stringify(intents, undefined, 2)
+export const safeId = (length = 10) => nanoid('1234567890abcdefghijklmnopqrsuvwxyz', length)
+
+export const makeId = (item: Item) => {
+  const firstQuestion = item.questions[Object.keys(item.questions)[0]][0]
+  return `__qna__${safeId()}_${slugify(firstQuestion)
+    .replace(/^_+/, '')
+    .substring(0, 50)
+    .replace(/_+$/, '')}`
+}
+
+export const normalizeQuestions = (questions: string[]) =>
+  questions
+    .map(q =>
+      q
+        .replace(/[\r\n]+/g, ' ')
+        .replace(/\s+/g, ' ')
+        .trim()
+    )
+    .filter(Boolean)
 
 export const isQnaComplete = (qnaEntry: Item, lang: string): boolean => {
   return qnaEntry.questions[lang]?.length >= QNA_MIN_QUESTIONS && qnaEntry.answers[lang]?.length >= QNA_MIN_ANSWERS
